+++ src/crosschain-liquidity/socket/SocketValidator.sol
@@ -1,6 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
 pragma solidity 0.8.19;
-
 import {MultiVaultsSFData, SingleVaultSFData} from "../../types/DataTypes.sol";
 import {BridgeValidator} from "../BridgeValidator.sol";
 import {ISuperRegistry} from "../../interfaces/ISuperRegistry.sol";
@@ -13,12 +12,12 @@ import "../../utils/DataPacking.sol";
 /// @author Zeropoint Labs
 /// @dev To assert input txData is valid
 contract SocketValidator is BridgeValidator {
-    mapping(uint16 => uint256) public socketChainId;
+    mapping(uint64 => uint256) public socketChainId;
 
     /*///////////////////////////////////////////////////////////////
                                 Events
     //////////////////////////////////////////////////////////////*/
-    event ChainIdSet(uint16 superChainId, uint256 socketChainId);
+    event ChainIdSet(uint64 superChainId, uint256 socketChainId);
 
     /*///////////////////////////////////////////////////////////////
                                 Constructor
@@ -42,8 +41,8 @@ contract SocketValidator is BridgeValidator {
     /// @inheritdoc BridgeValidator
     function validateTxData(
         bytes calldata txData_,
-        uint16 srcChainId_,
-        uint16 dstChainId_,
+        uint64 srcChainId_,
+        uint64 dstChainId_,
         bool deposit_,
         address superForm_,
         address srcSender_,
@@ -84,9 +83,9 @@ contract SocketValidator is BridgeValidator {
     /// @dev allows admin to add new chain ids in future
     /// @param superChainIds_ is the identifier of the chain within superform protocol
     /// @param socketChainIds_ is the identifier of the chain given by the bridge
-    function setChainIds(uint16[] memory superChainIds_, uint256[] memory socketChainIds_) external onlyProtocolAdmin {
+    function setChainIds(uint64[] memory superChainIds_, uint256[] memory socketChainIds_) external onlyProtocolAdmin {
         for (uint256 i = 0; i < superChainIds_.length; i++) {
-            uint16 superChainIdT = superChainIds_[i];
+            uint64 superChainIdT = superChainIds_[i];
             uint256 socketChainIdT = socketChainIds_[i];
             if (superChainIdT == 0 || socketChainIdT == 0) {
                 revert Error.INVALID_CHAIN_ID();
